requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-huggingface
chromadb
fastapi
uvicorn
python-multipart
streamlit
requests
sentence-transformers
torch
python-dotenv
rank_bm25
